Move polyfill stubs into polyfills/stubs and update imports
