+++ build-system/tasks/dep-check.js
@@ -153,7 +153,10 @@ Rule.prototype.matchBadDeps = function (moduleName, deps) {
           }
         }
         mustNotDependErrors.push(
-          cyan(moduleName) + ' must not depend on ' + cyan(dep)
+          cyan(moduleName) +
+            `[${this.config_.filesMatching} X ${this.config_.mustNotDependOn}]` +
+            ' must not depend on ' +
+            cyan(dep)
         );
       }
     });
