+++ extensions/amp-intersection-observer-polyfill/0.1/amp-intersection-observer-polyfill.js
@@ -17,7 +17,7 @@
 import {installIntersectionObserver} from 'intersection-observer/intersection-observer.install';
 import {maybeSetupCrossOriginObserver} from './cross-origin-observer';
 import {registerServiceBuilder} from '../../../src/service';
-import {upgradePolyfill} from '../../../src/polyfillstub/intersection-observer-stub';
+import {upgradePolyfill} from '../../../src/polyfills/stubs/intersection-observer-stub';
 
 const TAG = 'amp-intersection-observer-polyfill';
