+++ extensions/amp-resize-observer-polyfill/0.1/amp-resize-observer-polyfill.js
@@ -16,7 +16,7 @@
 
 import {installResizeObserver} from '../../../third_party/resize-observer-polyfill/ResizeObserver.install';
 import {registerServiceBuilder} from '../../../src/service';
-import {upgradePolyfill} from '../../../src/polyfillstub/resize-observer-stub';
+import {upgradePolyfill} from '../../../src/polyfills/stubs/resize-observer-stub';
 
 const TAG = 'amp-resize-observer-polyfill';
