+++ extensions/amp-resize-observer-polyfill/0.1/test/test-amp-resize-observer-polyfill.js
@@ -17,7 +17,7 @@
 import {
   ResizeObserverStub,
   installStub,
-} from '../../../../src/polyfillstub/resize-observer-stub';
+} from '../../../../src/polyfills/stubs/resize-observer-stub';
 import {upgradeResizeObserverPolyfill} from '../amp-resize-observer-polyfill';
 
 describes.realWin('amp-resize-observer-polyfill', {amp: false}, (env) => {
